fix: move generateEndpoint to new utils file to avoid loading modules unavailable in middleware
